feat(nasa): add endpoint to get asset by nasaId
Add a new endpoint to the NasaController to get an asset by its nasaId.
The endpoint is a GET request to '/nasa/asset/:nasaId' and it returns a
NasaAsset object.

Also, the NasaAsset model was added to the project to support the new
endpoint.

Finally, a new request was added to the thunderclient.json file to test
the new endpoint.